♻️ create dark mode hooks
